# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lotus Driver Tracking System",
    "description": "Backend for driver signup/login, barcode-scan order registration and delivery completion (GPS + timestamp logged per scan), and an SSO-gated admin dashboard for reviewing driver activity.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}/events": {
      "get": {
        "summary": "List the chronological scan events (registered, delivered) for one of the current driver's jobs.",
        "operationId": "listJobEvents",
        "tags": ["driver"],
        "security": [{ "driverBearer": [] }],
        "parameters": [{ "name": "job_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Event list, oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } }
                }
              }
            }
          },
          "404": { "description": "Job not found." }
        }
      }
    },
...
    "/api/admin/drivers": {
      "get": {
        "summary": "List all drivers, most recently registered first.",
        "operationId": "adminListDrivers",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "responses": {
          "200": {
            "description": "Driver list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "drivers": { "type": "array", "items": { "$ref": "#/components/schemas/Driver" } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/warehouses": {
      "get": {
        "summary": "List every warehouse outlet, including removed (inactive) ones.",
        "operationId": "adminListWarehouses",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "responses": {
          "200": {
            "description": "Outlet list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "warehouses": { "type": "array", "items": { "$ref": "#/components/schemas/Warehouse" } } }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new warehouse outlet.",
        "operationId": "adminCreateWarehouse",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WarehouseRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "warehouse": { "$ref": "#/components/schemas/Warehouse" } } }
              }
            }
          },
          "422": { "description": "Name is empty." }
        }
      }
    },
    "/api/admin/warehouses/{warehouse_id}": {
      "put": {
        "summary": "Edit a warehouse outlet's name/address.",
        "operationId": "adminUpdateWarehouse",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [{ "name": "warehouse_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WarehouseRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "warehouse": { "$ref": "#/components/schemas/Warehouse" } } }
              }
            }
          },
          "404": { "description": "Outlet not found." },
          "422": { "description": "Name is empty." }
        }
      },
      "delete": {
        "summary": "Remove a warehouse outlet (soft-delete: marks it inactive rather than deleting the row, since drivers may already be assigned to it).",
        "operationId": "adminRemoveWarehouse",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [{ "name": "warehouse_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Removed.",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } }
          },
          "404": { "description": "Outlet not found." }
        }
      }
    },
    "/api/admin/jobs": {
      "get": {
        "summary": "List delivery jobs across all drivers, filterable by status, driver, session and date range.",
        "operationId": "adminListJobs",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "driver_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "manifest_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated job list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/AdminJob" } },
                    "page": { "type": "integer" },
                    "page_size": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/exports/jobs.csv": {
      "get": {
        "summary": "Download the full job/event log as CSV -- one row per scan event (registered/delivered/failed, with GPS, failure reason, proof-photo URL), not just each job's current status. Same filters as GET /api/admin/jobs, unpaginated.",
        "operationId": "adminExportJobsCsv",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "driver_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "manifest_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "CSV file download (Content-Disposition: attachment).",
            "content": { "text/csv": { "schema": { "type": "string" } } }
          }
        }
      }
    },
    "/api/admin/jobs/{job_id}": {
      "get": {
        "summary": "Get full detail for one delivery job, including its session and driver.",
        "operationId": "adminGetJob",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [{ "name": "job_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Job detail.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "job": { "$ref": "#/components/schemas/AdminJobDetail" } } }
              }
            }
          },
          "404": { "description": "Job not found." }
        }
      }
    },
    "/api/admin/jobs/{job_id}/events": {
      "get": {
        "summary": "Get the full chronological scan-event log for a job (timestamp, status, GPS).",
        "operationId": "adminGetJobEvents",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "parameters": [{ "name": "job_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Event list, oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "events": { "type": "array", "items": { "$ref": "#/components/schemas/AdminEvent" } } }
                }
              }
            }
          },
          "404": { "description": "Job not found." }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List admin users on the SSO allowlist, oldest first.",
        "operationId": "adminListAdmins",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "responses": {
          "200": {
            "description": "Admin list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "admins": { "type": "array", "items": { "$ref": "#/components/schemas/AdminUser" } } }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new admin user to the SSO allowlist.",
        "operationId": "adminAddAdmin",
        "tags": ["admin"],
        "security": [{ "adminSSO": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddAdminRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Admin added.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "id": { "type": "integer" }, "email": { "type": "string" }, "name": { "type": "string" } }
                }
              }
            }
          },
          "409": { "description": "A user with this email already exists." }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "driverBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Driver JWT from /api/auth/signup or /api/auth/login, signed with the platform's JWT_SECRET."
      },
      "adminSSO": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Forwarded-Email",
        "description": "Injected by the Substrait platform's Google SSO proxy when SSO is enabled for this app in the portal's Access tab; not client-supplied. The caller must also be on the admin allowlist in the users table."
      }
    },
    "schemas": {
...
      "AddAdminRequest": {
        "type": "object",
        "required": ["email", "name"],
        "properties": { "email": { "type": "string" }, "name": { "type": "string" } }
      },
...
      "WarehouseRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "address": { "type": "string", "nullable": true }
        }
      },
...
      "Warehouse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "address": { "type": "string", "nullable": true },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string" }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tracking_no": { "type": "string" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed", "cancelled"] },
          "created_at": { "type": "string" }
        }
      },
      "AdminJob": {
        "allOf": [
          { "$ref": "#/components/schemas/Job" },
          {
            "type": "object",
            "properties": {
              "manifest_id": { "type": "integer" },
              "work_date": { "type": "string", "format": "date" },
              "warehouse_arrived_at": { "type": "string", "nullable": true },
              "driver_id": { "type": "integer" },
              "driver_name": { "type": "string" }
            }
          }
        ]
      },
      "AdminJobDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tracking_no": { "type": "string" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed", "cancelled"] },
          "created_at": { "type": "string" },
          "manifest_id": { "type": "integer" },
          "work_date": { "type": "string", "format": "date" },
          "warehouse_arrived_at": { "type": "string", "nullable": true },
          "driver_id": { "type": "integer" },
          "driver_name": { "type": "string" },
          "driver_email": { "type": "string" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed"] },
          "occurred_at": { "type": "string" },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "failure_reason": { "type": "string", "nullable": true },
          "photo_id": { "type": "integer", "nullable": true }
        }
      },
      "AdminEvent": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "job_id": { "type": "integer" },
          "driver_id": { "type": "integer" },
          "status_code": { "type": "string", "enum": ["registered", "delivered", "failed"] },
          "occurred_at": { "type": "string" },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "failure_reason": { "type": "string", "nullable": true },
          "photo_id": { "type": "integer", "nullable": true }
        }
      },
...
      "Driver": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "warehouse_id": { "type": "integer", "nullable": true },
          "warehouse_name": { "type": "string", "nullable": true },
          "created_at": { "type": "string" }
        }
      },
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "created_at": { "type": "string" }
        }
      }
    }
  }
}